# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Security-first OpenAPI catalog API for SpecRelay
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List workspaces
      responses:
        "200":
          description: Workspace list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceInput"
      responses:
        "201":
          description: Created workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "400":
          $ref: "#/components/responses/BadRequest"
  /workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get a workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: MCP-Protocol-Version
          in: header
          required: true
          schema:
            type: string
            const: "2026-07-28"
        - name: Mcp-Method
          in: header
          required: true
          schema:
            type: string
            enum: [server/discover, tools/list, tools/call]
        - name: Mcp-Name
          in: header
          required: false
          description: Required for tools/call and must match params.name.
          schema:
            type: string
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "404":
          $ref: "#/components/responses/NotFound"
  /workspaces/{workspaceId}/overview:
    get:
      operationId: getWorkspaceOverview
      tags: [workspaces]
      summary: Get workspace catalog overview
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Catalog overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceOverview"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /workspaces/{workspaceId}/mcp:
    post:
      operationId: invokeMcp
      tags: [mcp]
      summary: Discover or invoke approved MCP tools
      description: >
        Stateless MCP 2026-07-28 Streamable HTTP endpoint supporting
        server/discover, tools/list, and tools/call. Requests require the
        protocol's per-request metadata and matching MCP-Protocol-Version,
        Mcp-Method, and, for tools/call, Mcp-Name headers. Only
        explicitly approved unauthenticated HTTPS GET operations from the
        latest stored specification are discoverable and executable.
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/McpRequest"
      responses:
        "200":
          description: JSON-RPC result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Operation is not approved for execution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Upstream request failed or was denied by the broker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: Upstream request timed out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: array
          items:
            type: string
      required: [error]
      additionalProperties: false
    Workspace:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, createdAt, updatedAt]
      additionalProperties: false
    WorkspaceInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
      required: [name]
      additionalProperties: false
    WorkspaceOverview:
      type: object
      properties:
        workspace:
          $ref: "#/components/schemas/Workspace"
        role:
          type: string
          enum: [OWNER, MEMBER]
        canManage:
          type: boolean
        apiCount:
          type: integer
          minimum: 0
        operationCount:
          type: integer
          minimum: 0
        enabledOperationCount:
          type: integer
          minimum: 0
        recentAuditEvents:
          type: array
          items:
            $ref: "#/components/schemas/AuditEvent"
      required: [workspace, role, canManage, apiCount, operationCount, enabledOperationCount, recentAuditEvents]
      additionalProperties: false
...
    AuditEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        eventType:
          type: string
        resourceType:
          type: [string, "null"]
        resourceId:
          type: [string, "null"]
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
      required: [id, workspaceId, eventType, resourceType, resourceId, metadata, createdAt]
      additionalProperties: false
    McpRequest:
      type: object
      properties:
        jsonrpc:
          type: string
          const: "2.0"
        id:
          anyOf:
            - type: string
            - type: integer
            - type: "null"
        method:
          type: string
          enum: [server/discover, tools/list, tools/call]
        params:
          type: object
          additionalProperties: true
      required: [jsonrpc, id, method, params]
      additionalProperties: false
    McpResponse:
      type: object
      description: JSON-RPC response containing either a tool list or tool call result.
      additionalProperties: true